Fix plugin forwards and backwards compatibility (#23)
Use the parcel plugin config symbol. This lets parcel plugins work with this tool and will allow all plugins from this tool to work with parcel.

In the future the atlaspack plugins can skip this symbol and just pass around objects that implement an interface. Adding the symbol into the plugin would be a compatibility layer.

## Motivation

Fix parcel plugin compatibility

* It'd be possible to handle both atlaspack and parcel symbols here, but that will mean plugins declared with `@atlaspack/plugin` will not work with parcel