Add faster packager feature-flag
This change adds a fast-path to `ScopeHoistingPackager::needsDefaultInterop`.

We have found that up to 50% of the build time can be spent on this function.
If we force the default interop to be true, we can skip this function entirely.

The impact on production bundles from real applications seems minimal and the
build time improvement is very significant.