Clean-up owned resolver feature-flags
We have found that `dashmap` was responsible for a majority of memory
crashes we were experiencing in our development environments. These were
very frequent and point at soundness issues in the library.

The new resolver is simplified and has no memory crashes. It is however
slightly slower. On large code-bases, both resolvers are not a bottleneck
and can resolve modules at more than sufficient speed.