Make changes to have all tests run always
This changes the v2/v3 test setup so that tests that are skipped
for V3 run. A subsequent change will improve our logging to
automatically diff the outputs of both versions in a single
process.

The idea here is to easily identify tests that are green but
skipped, as well as get a grasp of current failures.