Start work on Rust HTML transformer (#77)
There are 2 changes:

* Add rust HTML transformer and start to write test-suite for JavaScript HTML transformer
* Fix asset ID alias

HTML transformer implementation is not complete. I chose to use Servo's HTML parser. There's just a very barebones set-up at the moment.

Please see the commit messages for justifications. I'm removing the "non zero u32" alias because it's unused and incompatible with the current code. So I'd rather alias to string until we port all code to use u32.

We shouldn't have an alias that can't be used on the current code. A change to make the IDs u32 should be applied globally to everything.