Fix bug in hashing of rust
This is an example of why we should avoid doing manual implementations
of standard library traits.

Here Hash and PartialEq were manually implemented and then Eq derived.

This leads to incorrect code being generated by the compiler because it
will make unsafe optimisations that assume equality has a few properties
that are broken by a custom implementation.