Optimize request tracker cache invalidation
This commit fixes several performance problems in the cache invalidation
implementation. Overall, this improves performance of handling a large
number of events from updating `node_modules` on a large application by
1500x.

The request graph stores a tasks, and their inter-dependencies, as well
as their relationship with files. Relationship with files is used for
cache invalidation.

There are a few problems with the current implementation:

* Cache invalidation performs `O(events * tasks)` amount of work. This
  is not fixed
* The "file above" invalidation type has extremely poor performance
  because even though the files are structured in a prefix tree sort of
  structure, the implementation is listing all incoming nodes and then
  iterating them to check for file-path matches. This means that for
  every package.json/index.js/common-file-path file change, we will
  perform N comparisons, where N is the number of directories in the
  entire project + dependencies that contains one of these files. On
  large projects, this number is in the hundreds of thousands, and since
  this is effectively `O(N ^ 2)` with respect of that number, this takes
  an unmanageable amount of time to complete.

The current time to process 300k events (which is the number of events)
triggered on a yarn install is over 2 hours. After this change, the time
becomes under 4 seconds. There are still significant optimizations that
can be made.

A better write-up of these changes and validation of their correctness
will be shared on a subsequent document.