Make resolver caches use thread-local storage only
On #10 we have removed `dashmap` from this project dependencies.

The reason for that is we were observing an alarming amount of crashes
related to memory bugs on that crate on developer machines. We have
confirmed that those bugs were in the `dashmap` crate and they stopped
happening after it was removed.

However, the solution put in place made resolution around 3-5x slower
than what it was before with `dashmap` caches. That is due to contention
on the `RwLock` hashmap used for resolver internal caches.

This change moves those caches to use a thread-local map, which achieves
close performance to the `dashmap` solution, around 30% slower.

The entire resolution task can be completed very quickly on a large
project (few seconds), hence we prefer to have a more stable
implementation for now.