Fix dev dep request missing 'invalidateOnCreate'
This commit fixes a long-standing bug in the dev dep request where the
'invalidateOnCreate' property was missing.

The reason this happened was due to an optimisation in the worker to
main-thread communication, which ommitted certain fields from the
DevDepRequest type when the values were thought to be unnecessary.

This commit fixes the issue by:

* Ensuring the value exists at a type-level more clearly

A test that reproduces the bug is added.