Fix dev dep request missing 'invalidateOnCreate'
This commit fixes a long-standing bug in the dev dep request where the
'invalidateOnCreate' property was missing.

The reason this happened was due to an optimisation in the worker to
main-thread communication, which ommitted certain fields from the
DevDepRequest type when the values were thought to be unnecessary.

This decision was based on a build cache present on the workers.
However, at times, this cache was also set on the main-thread.

These build caches were not being cleared on the main-thread, only on
the worker threads. This caused the main-thread to sometimes use stale
dev-dep requests from a previous build and then throw an exception
since those dev-dep requests would execute but be missing the field.

This commit fixes the issue by:

* Ensuring the value exists at a type-level more clearly
* Clearing the build cache on the main-thread between builds

Tests that reproduce this can be found on the experimental bundler
branch:

* https://github.com/atlassian-labs/atlaspack/compare/issue/dominators-bundler?expand=1