Fix PromiseQueue to remove unhandled rejections
The current `PromiseQueue` API is build such that the following code
will raise an unhandled rejection error:

```js
const queue = new PromiseQueue();
queue.add(() => Promise.reject(new Error('foo')));
```

For this reason, most of the codebase uses `PromiseQueue#add` with a
catch block around the promise, including its own tests.

As such, the correct usage of the API is:
```js
const queue = new PromiseQueue();
queue.add(() => Promise.reject(new Error('foo'))).catch((err) => {});
```

All the other usages, that haven't been modified by this commit, were
missing this handling and may lead to unhandled rejections crashing
threads/workers.

This commit changes the API so that this can't happen, the API can't
trigger unhandled exceptions anymore, and the promise is not returned.

The only case that did require a promise to be returned was the
`AssetGraphRequest` which is updated to use the new API.