Add dominance algorithm implementation
Implements "A simple, fast dominance algorithm", to find the immediate
dominators of all nodes in a graph.

- https://www.cs.tufts.edu/comp/150FP/archive/keith-cooper/dom14.pdf

This will be used for a new experimental bundler package to be added.
This adds the empty package.

Test Plan: Run `yarn test` to execute unit-tests for the implementation

Reviewers: MonicaOlejniczak

Reviewed By: MonicaOlejniczak

Pull Request: https://github.com/atlassian-labs/atlaspack/pull/326