Add Tarjan SCC algorithm implementation
Adds an implementation of Tarjan's algorithm to find strongly connected
components.

This will be used on the experimental bundler to simplify cycle handling and
can be used on other parts of the codebase to simplify the symbol propagation
implementation.

On the bundler, we will use this to convert the graph to an acyclic graph
(replace the strongly connected components with a single 'group' node) before
doing other work.

Test Plan: yarn test

Reviewers: JakeLane

Reviewed By: JakeLane

Pull Request: https://github.com/atlassian-labs/atlaspack/pull/332