Optimize resolver performance by reducing calls to Specifier::to_string
At the moment, for every import we will loop over the list of merged
`tsconfig.json` paths and convert the `Specifier`s on those paths into strings.

For all of the "package" specifiers, this involves a `format!` call, which is
taking quite a considerable amount of time to execute.

We can improve this by:

* caching the `Specifier::to_string` results for a tsconfig object, through the
  duration of the build

Test Plan: cargo test

Reviewers: JakeLane, MonicaOlejniczak

Reviewed By: JakeLane, MonicaOlejniczak

Pull Request: https://github.com/atlassian-labs/atlaspack/pull/376