Add multi-arch Node.js docker image
Built on ubuntu 22.04.

We have smoke tests for `armhf` and `arm64` which run on a docker image. This docker image is using an older version of GLIBC, which is incompatible with the version on `22.04`. Ubuntu `20.04` runners aren't working anymore, so the build runners will be upgraded to `22.04`.

This diff will add a build for docker images with Node.js for those architectures, based on ubuntu 22.04. We will then update the smoke test workflow to use these. I need to merge to be able to run the workflow.

Test Plan:
N/A

[no-changeset]: Adding docker images for smoke testing

Reviewers: JakeLane

Reviewed By: JakeLane

Pull Request: https://github.com/atlassian-labs/atlaspack/pull/447