Fix creating tag on GH canary releases
The current canary release job is flaky.

The reason is it is trying to create a git tag for the currently published NPM
package version right after publishing.

However, querying NPM after publish is not reliable because NPM is eventually
consistent.

This commit tries to change the canary job to allow lerna to commit the version
bump, then use the package.json information to find the published version.

[no-changeset]: Workflow changes

Test Plan: N/A

Reviewers: benjervis

Reviewed By: benjervis

Pull Request: https://github.com/atlassian-labs/atlaspack/pull/557