Fix LMDB database running out of reader slots
It is possible on certain use cases that the LMDB database will run out of
reader slots if processes are quit while keeping the database open.

This adds a call which will check and clean stale readers, fixing a database
that has become unable to be opened.

Test Plan: N/A

Reviewers: mattcompiles

Reviewed By: mattcompiles

Pull Request: https://github.com/atlassian-labs/atlaspack/pull/594