Add bundleId to write bundles request result
When building cache analysis tools, it is useful to be able to link a write
bundles request to the bundle that it has written, since the final output path
of a bundle is not available otherwise.

This modifies the WriteBundleRequest to include the bundleId that has been
written in the request result, which will get persisted to its request node on
the cache.

Test Plan: N/A

Reviewers: pancaspe87

Reviewed By: pancaspe87

Pull Request: https://github.com/atlassian-labs/atlaspack/pull/595