Reproduce content key not found errors
A common bug seems to be that the dev-server eventually hits exceptions:

```
"Expected content key ... to exist"
```

When this happens, the cache must be cleared, otherwise the dev-server is
unable to build.

Similar to #652, the reason this might happen is due to state corruption.

This PR adds a test which reproduces the exception by executing the following
sequence of events:

- Perform a build
- Perform a build which is not safe to incrementally bundle (such as adding a
  new dependency)
- "Abort" the build before bundling completes
- Perform a new build, which has changes to assets that haven't been seen by
  the bundler

This causes the asset graph to be marked as "safe to incrementally bundle",
despite the bundle graph being out-of-sync with it.

In general, removing the abort controller from the bundle graph request will
likely mitigate the issue.

However, the "safeToIncrementallyBundle" flag is slightly brittle as an
implementation here.

Ideally, our caches and state would not contain so much mutation, and the
decision of whether to skip work or re-use previous data would be purer and
moved to the bundling call-site.

As a quick fix, we can track bundling requirements differently.

For any asset graph, we will keep a marker indicating whether it needs to be
bundled. This will be an incrementing number. When we detect the graph requires
bundling, we will tick this number. Subsequent builds' asset graphs will
inherit this marker from the previous asset graph result.

When running the bundle graph request, the bundle graph will store the "asset
graph version for bundling" in itself. When triggering an incremental bundling
task, we will check whether the previous result's "asset graph version for
bundling" matches the current one.